Mini Player Wrapper Iteration 2:  Updated android build and moved imports inside src folder.
