+++ android/build.gradle
@@ -37,6 +37,6 @@ repositories {
 
 dependencies {
     implementation 'com.facebook.react:react-native:+'
-    implementation 'tv.vizbee:android-sender-sdk:6.5.7-persistent_mini_cast_bar-rc2'
+    implementation 'tv.vizbee:android-sender-sdk:6.5.7-persistent_mini_cast_bar-rc3'
 }
